Continue electron set-up
It's becoming apparent that, for an application whose front end can
make requests to the back end via both electron's IPC an axios calls
to a remote server, a common interface file supplied by the backend is
needed so that the front-end code can remain unchanged. Or, some
mechanism which allows calls to the backend to be fulfilled without
changing how the front-end calls it.