Read meta-data from WAV files.
Comments needed, but effectively uses the RIFF chunk building block to read information in about the WAV file.
Each RIFF chunk looks like the following
- Chunk ID
- Chunk size
- Chunk data

The entire WAV file is one main RIFF chunk, but the chunk data is composed of subchunks. The first subchunk
describes the format of the file, i.e., number of channels, sample rate, etc. This data has been
organised into a struct, so that it can be conveniently read straight in (see line `24` of `WAVHandler.cpp` in this commit.) The struct
is called `FMT_CK` for "format chunk".

Finally, following the format data, there is a second 'sub chunk' which contains the data of the file.

`loadMeta` handles the format chunk/the 8 extra bytes of the main chunk ID/size.
`load` will then handle the full reading of the file.